requirements.txt
appdirs==1.4.3
astroid==2.0.4
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
certifi==2019.6.16
cffi==1.12.3
Click==7.0
importlib-metadata==0.17
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
more-itertools==7.0.0
numpy==1.16.4
olefile==0.46
packaging==19.0
Pillow==6.1.0
pluggy==0.12.0
py==1.8.0
pycparser==2.19
pylint==2.1.1
pyparsing==2.4.0
pytest==5.0.1
PyYAML==5.1.1
rope==0.14.0
scipy==1.3.0
six==1.11.0
toml==0.10.0
torch==1.1.0
torchvision==0.3.0
wcwidth==0.1.7
wrapt==1.10.11
zipp==0.5.1
